Added USGS AirportPoint parsing
AirportPoint shapefile is converted externally with QGIS and saved as a csv file. Converting in QGIS was quicker than writing a function to pull the shapefile out of the archive and convert in python.

To support this, added a couple lat/lon functions that convert to degrees-minutes-seconds for future use (identifying which data sources had truncated or mis-transcribed data).