Added get match function
Match checks for proximity and name. This pass, the names must be similar to be counted as a match. 

Added source_id field to map back to the record in the source date. 

Added USGS enumeration for airport type. (Still no open/closed indicator for USGS except in runways file)